requirements.txt
openai
ell-ai
python-dotenv